Rename uma2 module to rpt because it only has the rpt functionality in it
